pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zo-dispatcher"
version = "1.0.0"
description = "A general-purpose agent dispatcher for Zo Computer that fires agents on schedules and webhook events."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "JackALaing" },
]
dependencies = [
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "requests>=2.31.0",
]

[project.scripts]
dispatcher-cli = "zo_dispatcher.cli:main"

[project.urls]
Homepage = "https://github.com/JackALaing/zo-dispatcher"

[tool.setuptools.packages.find]
include = ["zo_dispatcher*"]
